Introduce the client package
The client package provides services to the cmd package.  Previously,
the cmd package used to call directly into the ccvm package but this
is no longer appropriate as there will soon be an IPC barrier between
the cmd package and the ccvm package.  Going forward the client
package will be modified to communicate with the ccvm package over
rpc.  This commit insulates the cmd package from these changes.
