+++ ccvm/ccvm_test.go
@@ -100,7 +100,7 @@ func TestSystem(t *testing.T) {
 		t.Errorf("Failed to Restart instance: %v", err)
 	}
 
-	_, _, err = waitForSSH(ctx)
+	_, _, err = WaitForSSH(ctx)
 	if err != nil {
 		t.Errorf("Instance not available via SSH: %v", err)
 	}
