+++ cmd/start.go
@@ -19,7 +19,7 @@ package cmd
 import (
 	"flag"
 
-	"github.com/intel/ccloudvm/ccvm"
+	"github.com/intel/ccloudvm/client"
 	"github.com/intel/ccloudvm/types"
 	"github.com/spf13/cobra"
 )
@@ -35,7 +35,7 @@ var startCmd = &cobra.Command{
 		defer cancelFunc()
 
 		mergeVMOptions(&startSpec, &startMOptsSpec)
-		return ccvm.Start(ctx, &startSpec)
+		return client.Start(ctx, &startSpec)
 	},
 }
